Fixed network and node insertion to adapt to new code. Still need more testing
