+++ .gitignore
@@ -28,3 +28,4 @@ MANIFEST
 
 # UNA Test Data
 Cities/
+Test Cases/
