--- madina/zonal/np_network.py
@@ -1,1 +1,0 @@
-import numpy as np
